+++ ChangeLog.md
@@ -1,6 +1,7 @@
 ## change log ##
 
 * 2016-05-01: 0.3.1:
+  - capture all ontology metadata in OntologyVersion so it gets saved in the db
   - minor refactor: writeOntologyFile now returns all the ontology metadata, while the interim
     extractSomeProps uses that metadata to extract the special values that are going in the 
     OntologyVersion object (these special values will be removed later on)